+++ src/index.js
@@ -1,5 +1,5 @@
 import pack from './pack/index.js'
 import unpack from './unpack/index.js'
-
+import version from './_version.js'
 // export pack
-export { pack, unpack }
+export { pack, unpack, version }