+++ unpackLib/quick.test.js
@@ -1,5 +1,5 @@
-import unpack from './efrt-unpack.min.js'
-import test from 'tape'
+const unpack = require('./efrt-unpack.min.js')
+const test = require('tape')
 
 test('unpack works:', function (t) {
   const str = 'true¦denmark,o0scandanavia;h0ntar0;io'